Description of the task generated by ChatGPT 4
Prompting : 

Title : Using a generative AI to implement a script to analyse a biological dataset

I would like to write a practical exercise addressed to biologists and bioinformaticians.
The goal is to lead them to use a generative AI (different models will be proposed) in order to generate a script for analysing some biological dataset.

The basic tasks could include the following steps :

- loading a tabular dataset
- performing some basic analysis on the columns and rows
- generating some derived table (e.g. marginal statistics on columns and rows)
- exporting the result tables in tab-separated value files (TSV format)
- exporting the same result tables in excel files (xlsx format)
- generating some figures to catch the main properties of the data (e.g. boxplots, violin plots, pie charts, …)
- exporting these figures in png and pdf formats

Constraints:

- Since the goal is to investigate the potential help of generative AI, the biological dataset should be straighforward (a biologist or bioinformatician should understand it before having to read much biological context about the data
- Similarly, the analysis should be relatively simple conceptually (not involving highly specific methods like RNA-seq normalisation, differential analysis etc)
- Each participant can use a scripting language of their choice, but we will provide support in “old-fashion” R, tidy verse R, or python

Importantly, the exercise should not provide the solution. Instead it should explain them how to use an AI to generate the code that will solve the problem.

The expectations would be to generate a properly documented and efficient script, that could be easily used by another user in order to reproduce the results, or to analyse another dataset of the same type.

At the end of the exercise, the participants should evaluate the usefulness of the AI for programming. I would then ask them to  fill a survey with this kind of questions:  

- Did the AI succeed to produce a functional script ?
- Was it easy or did it cost a lot of dialog with the AI to reach a result ?
- Did you have to fix the code manually or did the AI (after some prompting rounds) produce the final script without manual intervention ?
- Is the final script properly documented ?
- Is the final script properly structured (e.g. does it contain separate functions) ?
- Does the result fit the expectation (exported derived tables, figures, …) ?

Could you generate

- a text that clearly describes the goals and procedure of the exercise according to the above sketch ?
- the instructions to start the work ?
- suggest additional evaluation criteria ?